refactor: split routes and reame schema filesto snake_case
