+++ src/siscan/classes/requisicao_exame_mamografia_rastreio.py
@@ -4,10 +4,10 @@ import logging
 
 from src.siscan.exception import CartaoSusNotFoundError
 from src.siscan.classes.requisicao_exame import RequisicaoExame
-from src.siscan.schema.RequisicaoMamografiaRastreamentoSchema import (
+from src.siscan.schema.requisicao_mamografia_rastreamento_schema import (
     RequisicaoMamografiaRastreamentoSchema,
 )
-from src.siscan.schema.RequisicaoMamografiaDiagnosticaSchema import (
+from src.siscan.schema.requisicao_mamografia_diagnostica_schema import (
     RequisicaoMamografiaDiagnosticaSchema,
 )
 from src.siscan.classes.requisicao_exame_mamografia import (
